Make progress double instead of float

+++ src/mixins/java/org/spongepowered/common/mixin/api/minecraft/server/level/ServerPlayerMixin_API.java
@@ -297,7 +297,7 @@ public abstract class ServerPlayerMixin_API extends PlayerMixin_API implements S
     }
 
     @Override
-    public void sendBlockProgress(final int x, final int y, final int z, final float progress) {
+    public void sendBlockProgress(final int x, final int y, final int z, final double progress) {
         if (progress < 0 || progress > 1) {
             throw new IllegalArgumentException("Progress must be between 0 and 1");
         }
